refactor: playerStat 을 redis 말고 Jpa 를 통해 mysql 에 저장하도록 변경

+++ src/main/java/com/ksy/fmrs/repository/Player/PlayerStatRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PlayerStatRepository extends JpaRepository<PlayerStat, Integer> {
+public interface PlayerStatRepository extends JpaRepository<PlayerStat, Long> {
 }
